+++ CHANGELOG.md
@@ -14,6 +14,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added the `pack` command which builds and prepares a publish-ready package of contracts' wrappers
 - Added support for wallet IDs in mnemonic provider. Environment variables `WALLET_ID` or `SUBWALLET_NUMBER` should be set, or a .env file with them must be present in order for it to be usable
 
+### Fixed
+
+- Fix address format in testnet
+
 ## [0.33.1] - 2025-05-16
 
 ### Fixed